CSSTUDIO-2023: move providers to dedicated folder
