+++ src/providers/index.js
@@ -1,0 +1,11 @@
+import MuiThemeProvider from "./MuiThemeProvider";
+import StyledComponentsThemeProvider from "./StyledComponentsThemeProvider";
+import LocalizationProvider from "./LocalizationProvider";
+import ReduxProvider from "./ReduxProvider";
+
+export {
+    MuiThemeProvider, 
+    StyledComponentsThemeProvider,
+    LocalizationProvider, 
+    ReduxProvider
+};
