+++ src/test-utils/wrappers.js
@@ -1,20 +1,15 @@
-import { Provider } from "react-redux";
-import { ThemeProvider } from "@mui/material";
-import { ThemeProvider as StyledComponentsThemeProvider } from "styled-components";
-import { LocalizationProvider } from "../config/localization";
-import GlobalStyle from "config/GlobalStyle";
+import { MuiThemeProvider, StyledComponentsThemeProvider, LocalizationProvider, ReduxProvider } from "providers";
 
-export const AppWrapper = ({store, theme, styledComponentsTheme, children}) => {
+export const AppWrapper = ({store, children}) => {
     return (
-        <Provider store={store}>
-            <ThemeProvider theme={theme}>
-                <StyledComponentsThemeProvider theme={styledComponentsTheme}>
-                    <GlobalStyle />
+        <ReduxProvider store={store}>
+            <MuiThemeProvider>
+                <StyledComponentsThemeProvider>
                     <LocalizationProvider>
                         {children}
                     </LocalizationProvider>
                 </StyledComponentsThemeProvider>
-            </ThemeProvider>
-        </Provider>
+            </MuiThemeProvider>
+        </ReduxProvider>
     )
 };
